+++ app/models/retweet.rb
@@ -1,25 +1,23 @@
 # frozen_string_literal: true
 
 class Retweet < ApplicationRecord
+  include Notifiable
   belongs_to :tweet, counter_cache: true
   belongs_to :user
 
   has_one :timeline, dependent: :destroy
 
-  # 通知モデルとのポリモーフィック
-  has_many :notifications, as: :notifiable, dependent: :destroy
-
   validates :user_id, uniqueness: { scope: :tweet_id }
-
-  after_create_commit :create_notifications
 
   private
 
-  def create_notifications
-    return unless tweet.user != user
+  # 自分自身へのアクションかどうかの判定
+  def check_create_notification
+    tweet.user == user
+  end
 
-    notification = Notification.create(notifiable: self, user: tweet.user)
-    notification.save!
-    NotificationMailer.notice_mail(notification).deliver_now
+  # 通知受信者の指定
+  def notification_recipient
+    tweet.user
   end
 end
